backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
pdf2image==1.16.3
openai==1.10.0
python-dotenv==1.0.0
